+++ src-docs/src/routes.js
@@ -29,6 +29,8 @@ import { ColorPaletteExample } from './views/color_palette/color_palette_example
 
 import { IsColorDarkExample } from './views/is_color_dark/is_color_dark_example';
 
+import { PrettyDurationExample } from './views/pretty_duration/pretty_duration_example';
+
 import { UtilityClassesExample } from './views/utility_classes/utility_classes_example';
 
 // Component examples
@@ -376,6 +378,7 @@ const navigation = [
       InnerTextExample,
       I18nExample,
       IsColorDarkExample,
+      PrettyDurationExample,
       MutationObserverExample,
       OutsideClickDetectorExample,
       PortalExample,